requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
flask>=2.0.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
tqdm>=4.62.0
segmentation-models-pytorch>=0.2.0
